refactor: improve naming, response handling & package structure
