+++ src/course/course.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CoursesService } from './courses.service';
-import { CoursesController } from './courses.controller';
+import { CoursesService } from './course.service';
+import { CoursesController } from './course.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CourseEntity } from './course.entity';
