+++ src/course/course.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CoursesService } from './courses.service';
+import { CoursesService } from './course.service';
 
 describe('CoursesService', () => {
   let service: CoursesService;
