+++ src/course/response/course.response.ts
@@ -2,7 +2,8 @@ export class CourseResponse {
     id: number;
     title: string;
     description: string;
-    level: string;
+    difficultyLevel: string;
+    price: number;
     createdAt: Date;
     updatedAt: Date;
 
@@ -11,7 +12,8 @@ export class CourseResponse {
         this.id = course.id;
         this.title = course.title;
         this.description = course.description;
-        this.level = course.level;
+        this.difficultyLevel = course.difficultyLevel;
+        this.price = course.price;
         this.createdAt = course.createdAt;
         this.updatedAt = course.updatedAt;
     }
